add加修改

+++ orianna-api/src/main/java/com/mhc/orianna/api/dto/AssetDTO.java
@@ -96,7 +96,7 @@ public class AssetDTO implements Serializable{
     /**
      * 领用人id
      */
-    private Integer authStaffId;
+    private Long authStaffId;
     /**
      * 领用人花名
      */
