+++ orianna-dal/src/main/java/com/mhc/orianna/dal/domain/Asset.java
@@ -119,7 +119,7 @@ public class Asset extends BaseModel {
      * 领用人id
      */
     @TableField("auth_staff_id")
-    private Integer authStaffId;
+    private Long authStaffId;
     /**
      * 领用人姓名
      */
@@ -314,11 +314,11 @@ public class Asset extends BaseModel {
         this.borrowApprovalNo = borrowApprovalNo;
     }
 
-    public Integer getAuthStaffId() {
+    public Long getAuthStaffId() {
         return authStaffId;
     }
 
-    public void setAuthStaffId(Integer authStaffId) {
+    public void setAuthStaffId(Long authStaffId) {
         this.authStaffId = authStaffId;
     }
